Fixed the linting issues. (#45)
